[C-2132] unlink is safe on ios but not android (#2852)

+++ apps/audius-client/packages/mobile/src/store/offline-downloads/sagas/downloadQueueSagas/workers/downloadCollectionWorker.ts
@@ -181,5 +181,7 @@ async function writeCollectionMetadata(collection: UserCollectionMetadata) {
 
 async function removeDownloadedCollection(collectionId: CollectionId) {
   const collectionDir = getLocalCollectionDir(collectionId.toString())
+  const exists = await RNFetchBlob.fs.exists(collectionDir)
+  if (!exists) return
   return await RNFetchBlob.fs.unlink(collectionDir)
 }
