+++ apps/audius-client/packages/mobile/src/store/offline-downloads/sagas/watchRemoveOfflineItems.ts
@@ -36,10 +36,16 @@ function* deleteItemsFromDisk(action: RemoveOfflineItemsAction) {
         getLocalCollectionDir,
         item.id.toString()
       )
-      yield* call(RNFetchBlob.fs.unlink, collectionDirectory)
+      const exists = yield* call(RNFetchBlob.fs.exists, collectionDirectory)
+      if (exists) {
+        yield* call(RNFetchBlob.fs.unlink, collectionDirectory)
+      }
     } else if (item.type === 'track' && !trackStatus[item.id]) {
       const trackDirectory = yield* call(getLocalTrackDir, item.id.toString())
-      yield* call(RNFetchBlob.fs.unlink, trackDirectory)
+      const exists = yield* call(RNFetchBlob.fs.exists, trackDirectory)
+      if (exists) {
+        yield* call(RNFetchBlob.fs.unlink, trackDirectory)
+      }
     }
   }
 }
